+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Use clap directly for command line parsing, which leads to a different
    `--help` output for the `--sort` and `--search-providers` options.
  * Bump MSRV to 1.57
+ * Switch from `ansi_term` to `termcolor`
 
 ### Breaking